add new test to see if Activity Steam is not responding
